+++ game/src/enemybullet.lua
@@ -1,47 +1,25 @@
 local EnemyBullet = Object:extend()
 
-function EnemyBullet:new(coords)
+function EnemyBullet:new(path)
     self.image = Images.enemybullet
     self.width = self.image:getWidth()
     self.height = self.image:getHeight()
-    self.x = coords[1] - self.width / 2
-    self.y = coords[2]
+    self.path = path
     self.gone = false
-    self.speedX = 0
-    self.speedY = 0
-    self.setTarget = true
-
-    if coords[3] then
-        self.speedX = coords[3]
-        self.setTarget = false
-    end
-    if coords[4] then
-        self.speedY = coords[4]
-        self.setTarget = false
-    end
-end
-
-function EnemyBullet:target(player)
-    local diffX = player.x - self.x
-    local diffY = player.y - self.y
-    local dist = math.sqrt(diffX * diffX + diffY * diffY)
-    self.speedX = 300 * diffX / dist
-    self.speedY = 300 * diffY / dist
 end
 
 function EnemyBullet:update(dt)
-    self.y = self.y + (self.speedY * dt)
-    self.x = self.x + (self.speedX * dt)
-    if self.x < 0
-        or self.x > WINDOW_WIDTH
-        or self.y < 0
-        or self.y > WINDOW_HEIGHT then
+    self.path:update(dt)
+    if self.path.x < 0
+        or self.path.x > WINDOW_WIDTH
+        or self.path.y < 0
+        or self.path.y > WINDOW_HEIGHT then
         self.gone = true
     end
 end
 
 function EnemyBullet:draw()
-    love.graphics.draw(self.image, self.x, self.y)
+    love.graphics.draw(self.image, self.path.x, self.path.y)
 end
 
 return EnemyBullet